Rearrange code for VS Code integration
